Hack: Allow *compilation* in 64-bit mode
This allows compilation in 64-bit mode. Don't expect this to work or do
anything at all, this is purely for compilation.

+++ src/openloco/audio/vehicle_channel.cpp
@@ -10,9 +10,9 @@ using namespace openloco::interop;
 static std::tuple<sound_id, channel_attributes> sub_48A590(const vehicle* v)
 {
     registers regs;
-    regs.esi = (int32_t)v;
+    regs.esi = (intptr_t)v;
     call(0x0048A590, regs);
-    return std::make_tuple<sound_id, channel_attributes>((sound_id)regs.eax, { regs.ecx, regs.edx, regs.ebx });
+    return std::make_tuple<sound_id, channel_attributes>((sound_id)regs.eax, { (int32_t)regs.ecx, (int32_t)regs.edx, (int32_t)regs.ebx });
 }
 
 vehicle_channel::vehicle_channel(channel&& c)
